[client] Truncate arguments before logging it
In case of API call exception the client will try to log the API function
name with the parameters. It is possible that one of the argument is too
large to log the full content of it (for example the 'b64zip' parameter of the
'massStoreRun' API function). For this reason we have to truncate the arguments.