[server] Fix getting analyzer statistics for runs
Previously all the statistics were queried for the given run ids and sometimes
not the latest one was shown on the GUI for a run. The reason was that no sorting
option was specified for the query and the order was random.
To solve this problem this patch rewrites the whole query part by using a subquery
and it will get statistics for the specific history ids.