[doc] Using details/summary in md files to show help message
The documentation files contain a lot command line help messages in raw
format and this way it is hard to read the docs. For this reason with this
patch we will use the HTML5 details feature which will collapse the help
messages by default and it will not take up so much space.