[test] Fix running web unit tests
The unit tests for the client and server codes are in separated directory.
Both directories contain a separate Makefile and we include these Makefiles
to a common Makefile in the web directory. Because both (server and client)
Makefiles used the same variable name (UNIT_TEST_CMD) when included in the
common Makefile the latest one was used and the unit tests for the client
part was run 2 times.
In this patch we will use different variable names for these tests and we
will fix the failed unit tests.