[server] Configure keepalive
When using a CodeChecker server with `Docker Swarm` on the server side
we may get a `[Errno 104] Connection reset by peer` exception when for
example the storage takes more than 15 minutes. With the keepalive
parameters we can solve this problem by not destroying the connection
when an operation takes too much time.