[server] Add more info log to the store API function
The storage API request (`massStoreRun`) is a very heavy operation and most
of the cases when the user stores huge amount of reports to the database it
takes too much time (10-20 minutes) to perform this operation. For this reason
in this patch we will insert some info log to see where the store operation
is.