[fix] Opt-in fix escape in logger in case of backslash, CR, LF
The build logger did not escape '\\', '\r' and '\n' correctly.
This patch fixes this issue while preserving the old version.

By default, the old version will be used, unless the
`CC_LOGGER_NEW_ESCAPING` environment variable is defined.

It is mentioned in the README.md.

I test both versions in the test files.
Also adds a few tips on how to debug such cases with confidence.