[fix] Handle None when printing a table/csv/rows
When printing a table (e.g.: `CodeChecker version`) it is possible
that a value is None (e.g.: a package is built from a release zip file
which will not contain git information). In this case when printing the
table the following exception will be raised:
```
unsupported format string passed to NoneType.__format__
```
In this patch we will convert None values to empty string to handle this
use case too.

We will also convert None values to a valid string value when printing
data in csv or in rows format.