[client] Handle suppression properly in diff command
Handle the use case when the diff command is called with 2 local report
directories and the source codes contain source code comments.

Also previously all the reports which contained any source code comment (even
if it was confirmed) were marked as suppressed during the diff command when
the basename or newname were a remote run. So if I introduced a new
bug with a confirmed source code comments and I run the diff command
with a remote run it doesn't show this report as new in the diff result.