[client] Fix source content change error when diffing remote runs
When diffing two runs from a command-line client, a
`The following source file contents changed since the latest analysis:`
error is shown instead of the diff.

To reproduce the problem we just have to remove the project folder after
storage or use the --trim-path-prefix option during storage and then
just run the diff command with two remote runs.