[analyze] Fix cleanup metadata
At the end of the analysis we try to cleanup the metadata file when we
do an incremental analysis. If the source file for a plist doesn't exist,
we try to remove the plist file from the system and from the metadata.

Previously we created a new metadata format but we forgot to modify this
code part where we remove the plist file from this metadata file.

This commit will handle the use case when old and new metadata format
already can be found in report directory during analysis and the
plist file need to be removed.