[clang-tidy] Allow to override checker list
If the file specified by the `--tidy-config` option contains a `Checks`
key or the `Checks` option is specified through the `--analyzer-config`
option the `-checks` option will be skipped from the actual clang-tidy
analyzer command, so only the checkers what specified in the `Checks`
key will be used on analysis.

If the `clang-tidy:take-config-from-directory=true` option is set, use
the `.clang-tidy` configuration files and skip setting the `-checks`
parameter of the clang-tidy binary.