[analyze] Fix ccache compiler detection
When we try to find compiler binaries from the `PATH` we will resolve the
symlinks by using `os.path.realpath` function. If the compiler
(`g++/gcc/clang`) is a symlink to `ccache` it will think that the compiler
is ccache. When we try to get for example the version of the detected
analyze it will return the version for the ccache binary and not for
the original compiler.
For this reason if the detected compiler binary is `ccache` we will use
the original compiler path instead of resolving the symlink.