[analyze] Fix duplication warning when creating failed zip
When there is a compilation failure, we will try to create a failed zip file
with source files. It is possible that we will try to add the same file
multiple times but on different path: `a/b/c` and `/a/b/../b/c`. In this
case CodeChecker will print out the following warning message:
`UserWarning: Duplicate name:...`
To avoid this problem we normalize the path before inserting it to the
zip file.