[client] Handle invalid proxy settings (#3198)
If `HTTP_PROXY` / `HTTPS_PROXY` environment variables are set and the
URL is invalid (e.g.: it doesn't contains a schema like localhost:8001)
the client will throw the following exception:
```
'NoneType' object has no attribute 'rfind'
  File "/codechecker/build/CodeChecker/cc_bin/CodeChecker.py", line 174, in <module>
    main(commands)
  File "/codechecker/build/CodeChecker/cc_bin/CodeChecker.py", line 141, in main
    sys.exit(args.func(args))
  File "//codechecker/build/CodeChecker/lib/python3/codechecker_client/cmd/store.py", line 856, in main
    traceback.print_stack()
```
From this error message it is hard to understand the real problem.

Unfortunately Thrift does not handle the use case when invalid proxy format
is used (e.g.: no schema is specified). For this reason we need to verify
the proxy format in our side.