[cmd] Introducing checker labels for profiles
This commit is the first step of a new approach of grouping checkers.
The plan is to select checkers by labels. Profiles, guidelines and
severities will be represented with labels instead of maintaining three
separate JSON files for these groupings.

In this first step profiles are listed in checker_labels.json and all
commands are getting this information from this file.