[report-converter][fix] Fix for the SpotBugs report converter
Some SpotBugs bug reports lack the 'start' and 'end' attributes in
their <SourceLine> tags. The converter must not crash in such cases
but instead take line number in the <SourceLine> tag in the innermost
block. If neither the blocks constin 'start' and 'end' attributes then
take 0 instead. This patch fixes this issue.