[test] Fix non existen report directory test
The `test_non_existent_reports_directory` test assumes that on the default
url (localhost:8001) doesn't run any server. Unfortunately for me a
CodeChecker server was started on this port and this test cases is failed
because the connection between the client and server was successfully
created. To solve this problem we should get a free port on the host
and we need to try to connect to this server. So this test case will
always pass.