[cmd] Checker name prefixes are meant along separator characters
Checkers can be enabled by groups, e.g. unix.cstring enables all
checkers starting with this prefix. However, unix.cs is not a checker
group so this shouldn't be allowed. In this case an error message is
printed and analysis doesn't start.

This commit contains some ugly implementations. This is not the final
form of this feature, a lot of refactoring will be done later. The
reason of the current solution is the approach of our next release.