[client] Fix storage of multiple report directory
It is possible that the same project is analyzed with different
configuration (e.g.: ctu, non-ctu) to different report directories and
the user want to store these results with one CodeChecker store command.

This command will create a storage zip file with all the plist files in it
under the reports directory. It is possible the the same file name can be
found in multiple report directories so for this reason we need to create a
unique plist file name when we build the zip file.