[web] Move generated API stubs to the repo
Modifying the Thrift API for external users is very hard. For this reason
in this patch we will modify the workflow and instead of pushing api packages
to pypi and npm we will create distributed packages from it (tar files) which
will be commited inside the repository. These packages will be installed by
pip and npm as local packages.