[server] Workaround for SQLite limitation in severity change
In SQLite there is a setting named SQLITE_LIMIT_COMPOUND_SELECT which
limits the maximal number of compound select statements. In changing
checker severities, we have such an SQL statement of which the size is
proportional to the number of checkers. If this limit is exceeded, then
the query fails. The default limit is 500, so the workaround is to
change severities in this big chunks.