[server] Fix failed files uniqueing on the statistics page
Unfortunately we can't distinct the failed file paths by using SQL
queries because the list of failed files for a run / analyzer are
stored in one column in a compressed way. For this reason we need to
decompress the value in the Python code before uniqueing and we can use
the `getFailedFiles` function in the `getFailedFilesCount`.