Not allowing disabling Clangsa modeling checkers
When a Clang Static Analyzer checker is disabled in CodeChecker,
clang is invoked with the "analyzer-disable-checker" flag.
This allows the user disabling core modeling checkers such as
unix.DynamicMemoryModeling. This causes malfunctioning of depending checkers.

Breaking Change:
After this patch, modeling and debug checkers (listed with clang -cc1 -analyzer-checker-help-developer)
will not be listed and cannot be disabled through CodeChecker with the --enable and --disable flags.

They can be enabled/disabled through the Clang Static Analyzer specific --saargs flag only.