[server] Fix getting existing analysis info
It is possible when multiple runs are stored simultaneously to the
server with the same analysis command that multiple entries are
stored into the database. In this case we will select the first one.

Using `one_or_none` sqlalchemy function is not an option because it
will raise an exception if multiple entries are exists in the database.