[cli] Log options in the configuration file
Multiple cli commands accepts a configuration file paramter (--config).
If someone is using it, from the logs we will not see the full command
what was executed, so it will be hard to start debugging the problem.
For this reason with this patch we will print out the extra options
from this configuration.