[server] Use processes instead of threads
In Python Threads are not very useful in our case because it can execute
only one statement at the same time even if multiple threads are started.
To be able to process multiple API requests simultatenously and use
the CPU cores of the machine we will use Processes when starting a
CodeChecker server.