[server] Unzip storage zip file to workspace directory
Previously the storage zip file is unzipped to the `/tmp` folder. It is
possible that the disk size of the host machine is not so big and when
the unzipped / decompressed zip size is too big that the user will
get `No space left on device` exception.

For this reason we will unzip the zip file to the workspace folder which
can be configured during a server start and it can point to an extra
network drive which size can be bigger than the host machine drive.

This patch also removes our `TemporaryDirectory` class because it is
available in the `tempfile` module since `Python 3.2`.