[cli] Do not print sensitive information when exception happens
It is possible that the `performLogin` API functions throws an exception.
On the client side we print the API function name and the argument list
if an exception happens. The argument list may contain sensitive
information such as password. In this patch we will skip printing
the argument list in this situation.