[cli] Handle use case when git is not available
It is possible that `git` is not available on the user's system where the
`CodeChecker store` command is executed. In this case importing git module
will fail. With this patch we will handle this use case and print a
warning message to the user.

This patch will also print some information to the user whether collecting
blame information was successfull or if no blame information is available
for source files.