[server] Fix getting git commit url
It is possible that a git commit url regex contains a group, which is
optional, so the value can be None. In this case when replacing the
group name with the value in the url will fail, because None can't be
used in replace. For this reason if the value is None, we will skip
it from the replacement.