[cmd] One analyzer for CheckerLabel methods
This is a refactoring commit. Some CheckerLabel methods can be given
analyzer names which filter the result to the checkers of the given
analyzers. This parameter was required to be an iterable, however,
string is also an iterable. If accidentaly an analyzer name was given
instead of a list of analyzers then it was also accepted mistakenly with
some strange interpretation. This is now fixed.