[gui] Fix console error on reports page
On the report filter bar for source component and cleanup plans filters
we show a pencil edit icon only if the current user has rights to manage
it. We set the `currentProduct` variable asynchronously, so it is possible
that it will be `null` when we try to get it by using the vuex getter method
and we will try to get the `administrating` property on a `null` value which
is of course throws an error.