[fix] Add static files to the pypi package
Previously in the `6.18.0` release `plist_to_html` package was removed and
the source code was moved to the `report_converter` module. Unfortunately we
forgot to update the `MANIFEST` file in the root dir of the repository,
so the pypi package isn't contained these static files. This commit will
solve this problem.