[server] Closed/open reports
An open or outstanding report reflects on a bug which should be fixed.

If a report is marked as false positive or intentional, then those are
not open reports because they don't need investigation.

When a report is marked as false positive or intentional, then its
fixed_at date should be set, because we consider these as closed.
"open report date" filter member selects reports which are not fixed yet
or their fix date is later than the current time. By setting fixed_at
column when maring a report as false positive makes these reports as not
outstanding ones.