[cli] Print full extended command when config file is used
If users are using CodeChecker config files and calls CodeChecker
with configuration filem from the client log we do not see the full
command which will be run with all the extra arguments where environment
variables are also expanded. For this reason we will log this command
to as an extra info message.