[gui] Fix number of outstanding reports chart
Let's assume that a bug was detected yesterday at 3PM and I want to view
this chart today at 1PM. In day resolution we used the current date as a
starting point and based on the interval length (e.g.: 7 days) we subtract
x days from the current date. For this reason the grap will not show me this
bug which was detected yesteday, but it assumes that is was detected today.

With this patch we will set the starting date to the end of the current date
which will solve this issue.