[test] Fix non-deterministic test in plist to html
In one test case related to plist to html conversion we iterate
over multiple plist files, we convert it to HTML files and we also
skip duplicates.

So let's assume that I have two plist files a.plist and b.plist and both
plist files contains the same results. Let's assume that I process the a.plist
file first and convert the reports to HTML. Next if I process the b.plist I will
do not generate HTML file from it because the same reports are already in the HTML
file what we generated for a.plist and we do not generate empty HTML files anymore.

To solve this problem in the test case we will check if we have report, and in this
case the HTML file has to be exist otherwise it hasn't.