[server][gui] Escape values for v-html attributes
We are using `v-html` attribute on the UI side to dinamically rendering
comments and analyzer commands. This can be very dangerous because it can
easily lead to XSS vulnerabilities. To solve this problem the server
will always return the escaped version of these values which can be safely
rendered on the UI.