[cli] Add `print-steps` option to `CodeChecker cmd diff`
Without bug steps it is hard to understood the problem by a programmer.
With this commit we will introduce a new option for the diff command
which can be used to print bug steps similar what we are doing at the
`CodeChecker parse` command.

This patch also solve the problem to print bug steps in HTML files for
reports which comes from a CodeChecker server.