[cli] Fix skipping reports
If a skip file was given to the `CodeChecker parse` command and one of
the mentioned file from the bug path was on the skip list we skipped that
report.

Example:
`main.cpp`:
```cpp
\#include "lib.h"

int main() {
  return foo(0);
}
```

`lib.h`:
```
double foo(int param) {
  return 1 / param;
}

```

`skip file content`:
```txt
+*/lib.h
-*
```

In this case if we analyzed this project and run on of the following
command, it returned with no results:
- `CodeChecker parse ./reports --skip skipfile.txt`
- `CodeChecker parse ./reports --file "*/lib.h"`

With this patch we will skip a report only if the last bug step is on the
skip list.