[server] Fix suppressing bug on the server
It is possible that the start line and end line of a bug reported by the
analyzer are different.

Example:
```cpp
int main()
{
  // codechecker_suppress [all] This will not be suppressed because of a bug.
  int x = 1 /
    0;
  return x;
}
```

In the above example the bug start line is where `x` variable is declared and
the end line is in the next line.

On the server we stored the start line as the line where the report can be
found but for suppression we checked whether there is a source code comment
above the end line. So if someone inserted a source code comment above the
start line the server hasn't suppressed that report.

With this patch we will solve this problem and we will check the source code
comment above the start line and not the end line.