[cli] Fix source line / file for remote reports
Previously if a run is stored to a CodeChecker server where the file paths are
trimmed and this run is given to `CodeChecker cmd diff` command, and
a remote report was in the result set of this command, the plain text output contained
an exception that the client failed to open the source file for reading. This
problem can be easily reproduced with the following steps:
 - Log and analyze a project.
 - Store the same results to two remote runs and use the `--trim-path-prefix` option.
 - Run the following command: `CodeChecker cmd diff -b remote_run1 -n remote_run2 --unresolved`

With this patch if HTML output format is given to the CodeChecker diff command,
we will try to fetch source file contents from the server and use it instead of
trying to get it from the user's local system.

Also if different output format is given (e.g.: plaintext) for performance reason
we will not fetch the whole source file contents but only the necessary source
line conents.