[analyzer] Print info message about logger tool
If intercept-build is available on the user's machine we will use it,
otherwise we will use the CodeChecker ld-logger tool in Linux environment.
It is possible that on Linux `intercept-build` is intalled but from the
`CodeChecker log` info messages we do not see it which tool is used to create
the compilation database.

With this patch we will print information to the user which logger tool is
used to create the compilation database.