[html] Fix static HTML report files
Previously `id` attribute was not set for source files when generated static
HTML report files for example with the `CodeChecker parse` command.
The JavaScript part used this attributes but it was always `undefined`.
This way if a user opened a report which bug step referenced mulitple different
source files, changing between the bug steps didn't change the source file.