[cli] Fix link in gerrit output
If `CC_REPORT_URL` is defined and `gerrit` format is used at `CodeChecker parse`
or `CodeChecker cmd diff` commands, the output will contain the value of this
environment variable wrapped inside quotes. When this output is sent to gerrit,
it will convert URL links to HTML `a` tags. Unfortunately gerrit will think that
the ending quote is part of the URL, so it will not remove it. This way the
URL will be invalid.

To solve this problem in this patch we will remove the quotes around the URL link.