[fix] Quote command line segment using shlex
Simply joining command segments is not enough for acquiring a valid
command. One should escape the necessary characters prior to that.
Escaping can be done manually, but using shlex is the preferred way
of doing so.

This patch might not fix all of the issues.

There were multiple places where this was not done adequately:
 * at calling `intercept-build`
 * at parsing the `compile_commands.json`'s `arguments` field
 * at extending the command with the referred response file's content
 * at extending the command without response files
 * at the `libtest/codechecker.py/all_command()` helper function
 * at a bunch of other places where a command is being logged - but
   for brevity, I've just fixed the two most important places of this
   - `analyzer_base.py/analyze()`
   - `analysis_manager.py/handle_reproducer()`