[feat] Review status belongs to report
Earlier a review status belonged to a bug hash. This way the users could
set the review status of a report in another user's personal run if it
shared the same bug.

So now we store the review status to each report instead of bug hash.

When a review status is set to a bug which didn't have one from source
code comment then we store the information that this report got a new
review status on the server. Each time a new report is stored with the
same bug hash, it will get this previously set review status as a
default value.

In this version a database migration happens so the review status is
stored in "reports" table instead of "review_statuses". During the
migration entries are removed from "review_statuses" where all reports
with the same bug hash have review status given in the source code.

[server] Closed/open reports

An open or outstanding report reflects on a bug which should be fixed.

If a report is marked as false positive or intentional, then those are
not open reports because they don't need investigation.

When a report is marked as false positive or intentional, then its
fixed_at date should be set, because we consider these as closed.
"open report date" filter member selects reports which are not fixed yet
or their fix date is later than the current time. By setting fixed_at
column when maring a report as false positive makes these reports as not
outstanding ones.