[fix] Fix HTML generation for CodeChecker cmd diff command
If the diff contained reports from multiple source files (e.g.: a.cpp + b.cpp)
the `CodeChecker cmd diff` command in HTML format generated HTML files for
each source file but inserted the same list of reports in all of the HTML files.

This can be easily reproduced by creating two source files and insert a report in both
of the source file. Create a compilation database with 2 translation units, analyze
this project and use the diff command to generate HTML output files to a run which is
empty.

This patch will solve this problem and will insert only those reports to a generated
HTML files which are really related to that file.