[analyzer] Allow --file and skipfile option to be given together
The CodeChecker VSCodePlugin uses the `--file` parameter to analyze single files.
Large projects load in their configuration using the `--config` parameter and
if there is a `-i skipfile` given in the config, `CodeChecker analyze` call
drops an error.

This patch will allow `-i skipfile` and `--file` to be given together.