[feat] Suppression rules apply to local report dir in remote diff
Setting a suppression rule for a report hash affects the local results
in a local report directory when diffing them to a remote run. For
example when storing one report to a run and suppressing it in GUI, the
result of "CodeChecker cmd diff -b remote_run -n report_dir --new" is
an empty list even if "CodeChecker parse report_dir" shows a report and
"CodeChecker cmd results remote_run" doesn't.