[web] Fix exceptions during blame information storage
The python git library may throw not only `TypeError` exceptions but
`ValueError` exceptions as well. For example when the `HEAD` file exists
in the `.git` directory but the user who is running the `CodeChecker store`
command doesn't have permission to this file.

To solve this problem we will catch all the exceptions when getting the
tracking branch and remote url.