[ctu] Don't silently ignore exceptions thrown in pre-analysis
Exceptions thrown in a thread pool by default stay in the thread pool
(won't be re-raised at the call to map_async or similar functions).
However, you can query whether an exception occured with the help of
AsyncResult.successful().

This patch is simple -- don't pretend that pre-analysis exceptions can
be ignored, just because they run asyncronously, exceptions there are
just as severe as elsewhere.