[analyzer] Cppcheck support
This commit adds support for the Cppcheck analyzer.

The following two commits are also squashed into this:
* [test] Add new test cases for Cppcheck support
* [analyzer] Check version compatibility

This commit is the resurrection of works made by Márton Csordás.
The original pull request #2290 had three more refactoring related change
sets which were omitted in this. Some changes were added in compared to
the original commits:
* The Cppcheck report converter is used to get the results in the
result_handler.
* The hash calculation logic has been uplifted to the latest one.
* Some small small compatibility changes were added because of the
changes in the code base.

Co-authored-by: bruntib <bruntib@users.noreply.github.com>
Co-authored-by: vodorok <vodorok@gmail.com>