[bugfix] Crash with intercept-build based compilation database
When compilation database is generated by intercept build then a build
action in this JSON file contains "arguments" field instead of "command"
which is a list of command-line arguments instead of a single string.

CodeChecker supports only "command" field, so this transformation must
be done in "CodeChecker analyze" command in order to avoid failure
when building SkipHandler with "--file <header_file>" flag.

Fixes #3683