Import the suppressions per report
CodeChecker cmd suppress run_name -i <import_file>
will only import suppressions for the run indicated by run_name,
and not all reports.

Suppressions are imported as single report suppressions
instead of suppression rules.

This is the behaviour the user expects when run name is given
as a parameter to the command.

This patch also fixes the changeReviewStatus(report_id, status, message)
API function to change the review status of a single report only
instead of all reports with the same hash.