[CLI] Don't ignore compiler warnings, even if clangtidy:take-config-from-directory=true is specified
clangtidy:take-config-from-directory is an analyzer config that makes
clang-tidy get its arguments from a .clang-tidy file, and only from that
file. What this implies, is that all other options on the command line
for clang tidy will be ignored.

The problem is that this also ignores compiler warnings. Now, compiler
warnings can be emitted through clang-tidy's bug report generation
infrastructure, but are NOT clang-tidy checks. They an option to the the
*compiler*, not *clang-tidy*. For this reason, we should not ignore
them, however we invoke clang-tidy.