Raise an exception instead of silently failing on a broken diagtool
Especially if you build your own clang, you are very likely to forget
building the diagtool target.

I personally build my clang with the following command:

ninja -C build/ clang clang-tidy

This patch serves as a reminder not to forget the diagtool target
either. Its not a big deal if its not built, but some tests fail, and
the silent failure leads to a miserable debugging experience.