[feat] Find by files anywhere on bugpath
In the GUI the set of reports can be filtered by filename or source
component. However, these filters are concerning the last bug point,
i.e. one can list the set of reports ending in a specific file.

In this patch a new filter option has been introduced which returns all
reports where the file is involved at any part of the bug path.