[bugfix] Invalid SQL query when filter for file and source component.
The error happened when getFileCounts() function got a filter where both
"filepath" and "componentNames" fields were used.
When both filters are set then they both extend the WHERE clause of the
SQL query. The added clauses were related to File table so this file
table was joined twice to the SQL command which is an invalid querey.
The previous solution to this problem was partial, because even more
occurrences of "JOIN File" could appear in the query. Only one must stay.